+++ spec/deployment/releases_spec.rb
@@ -1,0 +1,23 @@
+require 'spec_helper'
+
+describe Statistrano::Deployment::Releases do
+
+  describe "#new" do
+    it "creates a deployment with a name" do
+      deployment = Statistrano::Deployment::Releases.new("name")
+      deployment.name.should == "name"
+    end
+
+    it "respects the default configs" do
+      deployment = Statistrano::Deployment::Releases.new("name")
+      deployment.config.release_count.should == 5
+    end
+
+    it "allows configs to be changed" do
+      deployment = Statistrano::Deployment::Releases.new("name")
+      deployment.config.release_count = 10
+      deployment.config.release_count.should == 10
+    end
+  end
+
+end