+++ spec/spec_helper.rb
@@ -1,0 +1,3 @@
+require 'rspec'
+require 'pry-debugger'
+require 'statistrano'